collector/requirements.txt
opencv-python==4.1.1.26
pulsar-client==2.4.1